Fix a use-after-free bug.
None of my Swift code was reference counting the Data objects which the images were built from (unsafely), which meant that trying to access underlying info about the NSImage caused big problems (when attempting to save to disk, for example).

Now this is fixed with a pretty terrible intermediate wrapper thingy.